libft rework

+++ ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: dabeloos <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/10/09 16:46:11 by dabeloos          #+#    #+#             */
-/*   Updated: 2018/10/09 16:46:46 by dabeloos         ###   ########.fr       */
+/*   Updated: 2018/10/19 15:55:35 by dabeloos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 char		*ft_itoa(int n)
 {
-	return (ft_itoa2(n));
+	return (ft_itoa2((long)n));
 }
